imlemented task_12 and made small refactoring
