+++ src/main/java/driver/BaseTest.java
@@ -4,6 +4,7 @@ import org.openqa.selenium.WebDriver;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Listeners;
+import testNG.Listener;
 
 import static driver.DriverCreation.getDriver;
 import static driver.DriverCreation.quitDriver;
