+++ src/main/resources/testRunners.task_12/task_12PropetriesExecutor.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+<suite name="task_12WorkingWithProperties">
+    <test name="WorkingWithProperties">
+        <classes>
+            <class name="applicationPropertiesTests.AppPropertiesLoginTest">
+            </class>
+        </classes>
+    </test>
+</suite>
