+++ src/main/resources/testRunners/task_10/task_10EnumFilterExecutor.xml
